Refactor burnable and freezable tests to use redeployed contracts
The tests for burnable and freezable contracts have been refactored to use newly deployed contracts. Additionally, these tests have been exported as functions to be used in other testing scenarios, specifically in the newly added stableCoinStudio test suite. This change improves the modularity and reusability of the burnable and freezable tests.

Signed-off-by: Marcos Serradilla Diez <marcos@io.builders>